Thoughtbubble & small bug fixes
Finally got around to creating the thoughtbubble. With this I also fixed
some minor cosmetic bugs with dialogbubbles. These include the height of
the bubble and the right-side spikes on the shoutbubble.

I've also tweaked the number of spikes on the left and right side of the
shoutbubble. All bug fixes are also applied to the new thoughtbubble.