stateVariable
Added a variable in DialogManager that is meant to work with a
dialog-type state. This variable will be false if there is no
dialogBubble open that the player can interact with and true otherwise.
This helps the rest of the code check if the game should enter a
specific dialog-type-state in which, for example, the game is paused to
focus on the dialog.

I also cleaned up my code by putting it in folders and removing old
obsolete code.